# src/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a DEIP server API.",
    "version": "1.0.0",
    "title": "DEIP",
    "contact": {
      "email": "yahor.tsaryk@deip.world"
    }
  },
...
  "paths": {
...
    "/api/v2/user/profile/{_id}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Find user profile",
        "description": "Return a profile of user",
        "operationId": "getUserProfile",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "_id",
            "in": "path",
            "description": "_id",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/UserProfile"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/v2/user/email/{email}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Find user",
        "description": "Return an user",
        "operationId": "getUserByEmail",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "email",
            "in": "path",
            "description": "email",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
    "/api/v2/users/listing": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Find users",
        "description": "Return an array of users",
        "operationId": "getUsersListing",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "status",
            "required": true,
            "type": "array",
            "collectionFormat": "multi"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v2/users/team/{teamId}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Find users",
        "description": "Return an array of users",
        "operationId": "getUsersByTeam",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "teamId",
            "in": "path",
            "description": "teamId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
    "/api/v2/users/portal/{portalId}": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Find users",
        "description": "Return an array of users",
        "operationId": "getUsersByPortal",
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "portalId",
            "in": "path",
            "description": "portalId",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/User"
              }
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "User not found"
          }
        },
        "security": [
          {
            "Bearer": []
          }
        ]
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "UserProfile": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "portalId": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "pubKey": {
          "type": "string"
        },
        "status": {
          "type": "number"
        },
        "teams": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "attributes": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "value": {},
              "attributeId": {
                "type": "string"
              }
            }
          }
        },
        "roles": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "role": {
                "type": "string"
              },
              "label": {
                "type": "string"
              },
              "teamId": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "User": {
      "type": "object",
      "properties": {
        "_id": {
          "type": "string"
        },
        "portalId": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "attributes": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "value": {},
              "attributeId": {
                "type": "string"
              }
            }
          }
        },
        "balances": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "account": {
                "type": "string"
              },
              "assetId": {
                "type": "string"
              },
              "symbol": {
                "type": "string"
              },
              "precision": {
                "type": "number"
              },
              "amount": {
                "type": "string"
              }
            }
          }
        },
        "modules": {},
        "roles": {
          "type": "array",
          "items": {
            "type": "object",
            "properties": {
              "role": {
                "type": "string"
              },
              "label": {
                "type": "string"
              },
              "teamId": {
                "type": "string"
              }
            }
          }
        },
        "pubKey": {
          "type": "string"
        },
        "status": {
          "type": "number"
        },
        "teams": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "createdAt": {
          "type": "string",
          "format": "date-time"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        },
        "metadataHash": {
          "type": "string"
        }
      }
    },
...
  }
}